+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Changelog
 
+## master
+
+* Add `EveOnline::ESI::PublicContracts`
+* Add `EveOnline::ESI::Models::PublicContract`
+
 ## v0.34.0
 
 * Rename `EveOnline::ESI::Models::Race#alliance_id` to `#faction_id`. [Fix esi issue 812](https://github.com/esi/esi-issues/issues/812).